Renamed fixture and tidy up
