+++ src/Entity/Dog.php
@@ -152,6 +152,7 @@ class Dog implements UserInterface, PasswordAuthenticatedUserInterface
     {
         if (!$this->posts->contains($post)) {
             $this->posts[] = $post;
+            $post->setDog($this);
         }
 
         return $this;
@@ -160,6 +161,7 @@ class Dog implements UserInterface, PasswordAuthenticatedUserInterface
     public function removePost(Post $post): self
     {
         if ($this->posts->removeElement($post)) {
+            $this->posts->removeElement($post);
             // set the owning side to null (unless already changed)
             if ($post->getDog() === $this) {
                 $post->setDog(null);
